% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_algebra.R
\name{approximately_solve}
\alias{approximately_solve}
\title{Solve linear system with fallback}
\usage{
approximately_solve(a, b)
}
\arguments{
\item{a}{Square matrix}

\item{b}{Right-hand side vector or matrix}
}
\value{
Solution x such that a \%*\% x = b
}
\description{
Solve linear system with fallback
}
\keyword{internal}
