% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blp_agents}
\alias{blp_agents}
\title{BLP (1995) Automobile Agent Data}
\format{
A data frame with 4000 rows and the following columns:
\describe{
\item{market_ids}{Market year}
\item{weights}{Integration weights}
\item{nodes0}{Integration node dimension 0}
\item{nodes1}{Integration node dimension 1}
\item{nodes2}{Integration node dimension 2}
\item{nodes3}{Integration node dimension 3}
\item{nodes4}{Integration node dimension 4}
\item{income}{Simulated income}
}
}
\source{
\url{https://pyblp.readthedocs.io/en/stable/}
}
\usage{
blp_agents
}
\description{
Agent-level data for the BLP (1995) automobile study. Contains integration
nodes, weights, and income demographics.
}
\references{
Berry, S., Levinsohn, J., & Pakes, A. (1995). Automobile Prices in Market
Equilibrium. \emph{Econometrica}, 63(4), 841-890.
}
\keyword{datasets}
