% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economy.R
\name{BLPEconomy}
\alias{BLPEconomy}
\title{BLP Economy Base Class}
\description{
Base class storing product/agent data and shared computations.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BLPEconomy-new}{\code{BLPEconomy$new()}}
\item \href{#method-BLPEconomy-get_market_data}{\code{BLPEconomy$get_market_data()}}
\item \href{#method-BLPEconomy-compute_logit_delta}{\code{BLPEconomy$compute_logit_delta()}}
\item \href{#method-BLPEconomy-print}{\code{BLPEconomy$print()}}
\item \href{#method-BLPEconomy-clone}{\code{BLPEconomy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPEconomy-new"></a>}}
\if{latex}{\out{\hypertarget{method-BLPEconomy-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPEconomy$new(
  product_formulations,
  product_data,
  agent_formulation = NULL,
  agent_data = NULL,
  integration = NULL,
  rc_types = NULL,
  epsilon_scale = 1,
  costs_type = "linear",
  add_exogenous = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPEconomy-get_market_data"></a>}}
\if{latex}{\out{\hypertarget{method-BLPEconomy-get_market_data}{}}}
\subsection{Method \code{get_market_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPEconomy$get_market_data(market_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPEconomy-compute_logit_delta"></a>}}
\if{latex}{\out{\hypertarget{method-BLPEconomy-compute_logit_delta}{}}}
\subsection{Method \code{compute_logit_delta()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPEconomy$compute_logit_delta(rho = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPEconomy-print"></a>}}
\if{latex}{\out{\hypertarget{method-BLPEconomy-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPEconomy$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPEconomy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BLPEconomy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPEconomy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
