% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulation.R
\name{BLPFormulation}
\alias{BLPFormulation}
\title{BLP Model Formulation}
\description{
Configuration for designing model matrices from R formulas.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BLPFormulation-new}{\code{BLPFormulation$new()}}
\item \href{#method-BLPFormulation-build_matrix}{\code{BLPFormulation$build_matrix()}}
\item \href{#method-BLPFormulation-get_names}{\code{BLPFormulation$get_names()}}
\item \href{#method-BLPFormulation-has_prices}{\code{BLPFormulation$has_prices()}}
\item \href{#method-BLPFormulation-has_shares}{\code{BLPFormulation$has_shares()}}
\item \href{#method-BLPFormulation-print}{\code{BLPFormulation$print()}}
\item \href{#method-BLPFormulation-clone}{\code{BLPFormulation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPFormulation-new"></a>}}
\if{latex}{\out{\hypertarget{method-BLPFormulation-new}{}}}
\subsection{Method \code{new()}}{
Create a new formulation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPFormulation$new(formula, absorb = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{formula}}{Formula string or R formula}

\item{\code{absorb}}{Optional formula for absorbed fixed effects}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPFormulation-build_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-BLPFormulation-build_matrix}{}}}
\subsection{Method \code{build_matrix()}}{
Build design matrix from data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPFormulation$build_matrix(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data frame}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Matrix with named columns
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPFormulation-get_names"></a>}}
\if{latex}{\out{\hypertarget{method-BLPFormulation-get_names}{}}}
\subsection{Method \code{get_names()}}{
Get column names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPFormulation$get_names()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPFormulation-has_prices"></a>}}
\if{latex}{\out{\hypertarget{method-BLPFormulation-has_prices}{}}}
\subsection{Method \code{has_prices()}}{
Check if prices are in the formulation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPFormulation$has_prices()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPFormulation-has_shares"></a>}}
\if{latex}{\out{\hypertarget{method-BLPFormulation-has_shares}{}}}
\subsection{Method \code{has_shares()}}{
Check if shares are in the formulation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPFormulation$has_shares()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPFormulation-print"></a>}}
\if{latex}{\out{\hypertarget{method-BLPFormulation-print}{}}}
\subsection{Method \code{print()}}{
Print formulation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPFormulation$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPFormulation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BLPFormulation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPFormulation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
