% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iteration.R
\name{BLPIteration}
\alias{BLPIteration}
\title{Iteration Configuration}
\description{
Configuration for the BLP contraction mapping fixed-point iteration.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BLPIteration-new}{\code{BLPIteration$new()}}
\item \href{#method-BLPIteration-iterate}{\code{BLPIteration$iterate()}}
\item \href{#method-BLPIteration-print}{\code{BLPIteration$print()}}
\item \href{#method-BLPIteration-clone}{\code{BLPIteration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPIteration-new"></a>}}
\if{latex}{\out{\hypertarget{method-BLPIteration-new}{}}}
\subsection{Method \code{new()}}{
Create iteration configuration
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPIteration$new(method = "squarem", method_options = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Character: "simple", "squarem", or "return"}

\item{\code{method_options}}{Named list of options}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPIteration-iterate"></a>}}
\if{latex}{\out{\hypertarget{method-BLPIteration-iterate}{}}}
\subsection{Method \code{iterate()}}{
Run fixed-point iteration
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPIteration$iterate(initial, contraction)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{initial}}{Starting values}

\item{\code{contraction}}{Function: x -> list(x_new, weights) or just x_new}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List with values, converged, iterations, evaluations
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPIteration-print"></a>}}
\if{latex}{\out{\hypertarget{method-BLPIteration-print}{}}}
\subsection{Method \code{print()}}{
Print iteration configuration
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPIteration$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPIteration-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BLPIteration-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPIteration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
