% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market.R
\name{BLPMarket}
\alias{BLPMarket}
\title{BLP Market Computations}
\description{
Market-level computations for BLP demand estimation.
All probability, share, delta, markup, and Jacobian computations happen here.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BLPMarket-new}{\code{BLPMarket$new()}}
\item \href{#method-BLPMarket-compute_random_coefficients}{\code{BLPMarket$compute_random_coefficients()}}
\item \href{#method-BLPMarket-compute_mu}{\code{BLPMarket$compute_mu()}}
\item \href{#method-BLPMarket-compute_probabilities}{\code{BLPMarket$compute_probabilities()}}
\item \href{#method-BLPMarket-compute_shares}{\code{BLPMarket$compute_shares()}}
\item \href{#method-BLPMarket-compute_delta}{\code{BLPMarket$compute_delta()}}
\item \href{#method-BLPMarket-compute_shares_by_xi_jacobian}{\code{BLPMarket$compute_shares_by_xi_jacobian()}}
\item \href{#method-BLPMarket-compute_shares_by_theta_jacobian}{\code{BLPMarket$compute_shares_by_theta_jacobian()}}
\item \href{#method-BLPMarket-compute_xi_by_theta_jacobian}{\code{BLPMarket$compute_xi_by_theta_jacobian()}}
\item \href{#method-BLPMarket-compute_eta}{\code{BLPMarket$compute_eta()}}
\item \href{#method-BLPMarket-compute_costs}{\code{BLPMarket$compute_costs()}}
\item \href{#method-BLPMarket-compute_equilibrium_prices}{\code{BLPMarket$compute_equilibrium_prices()}}
\item \href{#method-BLPMarket-compute_elasticities}{\code{BLPMarket$compute_elasticities()}}
\item \href{#method-BLPMarket-compute_diversion_ratios}{\code{BLPMarket$compute_diversion_ratios()}}
\item \href{#method-BLPMarket-compute_consumer_surplus}{\code{BLPMarket$compute_consumer_surplus()}}
\item \href{#method-BLPMarket-compute_hhi}{\code{BLPMarket$compute_hhi()}}
\item \href{#method-BLPMarket-compute_markups}{\code{BLPMarket$compute_markups()}}
\item \href{#method-BLPMarket-compute_profits}{\code{BLPMarket$compute_profits()}}
\item \href{#method-BLPMarket-clone}{\code{BLPMarket$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPMarket-new"></a>}}
\if{latex}{\out{\hypertarget{method-BLPMarket-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPMarket$new(
  products,
  agents,
  sigma = NULL,
  pi = NULL,
  rho = NULL,
  rc_types = NULL,
  epsilon_scale = 1,
  costs_type = "linear"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPMarket-compute_random_coefficients"></a>}}
\if{latex}{\out{\hypertarget{method-BLPMarket-compute_random_coefficients}{}}}
\subsection{Method \code{compute_random_coefficients()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPMarket$compute_random_coefficients()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPMarket-compute_mu"></a>}}
\if{latex}{\out{\hypertarget{method-BLPMarket-compute_mu}{}}}
\subsection{Method \code{compute_mu()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPMarket$compute_mu(coefficients = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPMarket-compute_probabilities"></a>}}
\if{latex}{\out{\hypertarget{method-BLPMarket-compute_probabilities}{}}}
\subsection{Method \code{compute_probabilities()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPMarket$compute_probabilities(delta, mu = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPMarket-compute_shares"></a>}}
\if{latex}{\out{\hypertarget{method-BLPMarket-compute_shares}{}}}
\subsection{Method \code{compute_shares()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPMarket$compute_shares(probabilities, weights = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPMarket-compute_delta"></a>}}
\if{latex}{\out{\hypertarget{method-BLPMarket-compute_delta}{}}}
\subsection{Method \code{compute_delta()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPMarket$compute_delta(initial_delta, iteration, fp_type = "safe_linear")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPMarket-compute_shares_by_xi_jacobian"></a>}}
\if{latex}{\out{\hypertarget{method-BLPMarket-compute_shares_by_xi_jacobian}{}}}
\subsection{Method \code{compute_shares_by_xi_jacobian()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPMarket$compute_shares_by_xi_jacobian(probabilities, conditionals = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPMarket-compute_shares_by_theta_jacobian"></a>}}
\if{latex}{\out{\hypertarget{method-BLPMarket-compute_shares_by_theta_jacobian}{}}}
\subsection{Method \code{compute_shares_by_theta_jacobian()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPMarket$compute_shares_by_theta_jacobian(
  probabilities,
  conditionals = NULL,
  sigma,
  pi_mat = NULL,
  rho = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPMarket-compute_xi_by_theta_jacobian"></a>}}
\if{latex}{\out{\hypertarget{method-BLPMarket-compute_xi_by_theta_jacobian}{}}}
\subsection{Method \code{compute_xi_by_theta_jacobian()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPMarket$compute_xi_by_theta_jacobian(probabilities, conditionals = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPMarket-compute_eta"></a>}}
\if{latex}{\out{\hypertarget{method-BLPMarket-compute_eta}{}}}
\subsection{Method \code{compute_eta()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPMarket$compute_eta(probabilities, ownership = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPMarket-compute_costs"></a>}}
\if{latex}{\out{\hypertarget{method-BLPMarket-compute_costs}{}}}
\subsection{Method \code{compute_costs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPMarket$compute_costs(probabilities, prices, ownership = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPMarket-compute_equilibrium_prices"></a>}}
\if{latex}{\out{\hypertarget{method-BLPMarket-compute_equilibrium_prices}{}}}
\subsection{Method \code{compute_equilibrium_prices()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPMarket$compute_equilibrium_prices(
  costs,
  iteration,
  ownership = NULL,
  prices_init = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPMarket-compute_elasticities"></a>}}
\if{latex}{\out{\hypertarget{method-BLPMarket-compute_elasticities}{}}}
\subsection{Method \code{compute_elasticities()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPMarket$compute_elasticities(probabilities, prices = NULL, name = "prices")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPMarket-compute_diversion_ratios"></a>}}
\if{latex}{\out{\hypertarget{method-BLPMarket-compute_diversion_ratios}{}}}
\subsection{Method \code{compute_diversion_ratios()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPMarket$compute_diversion_ratios(probabilities, name = "prices")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPMarket-compute_consumer_surplus"></a>}}
\if{latex}{\out{\hypertarget{method-BLPMarket-compute_consumer_surplus}{}}}
\subsection{Method \code{compute_consumer_surplus()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPMarket$compute_consumer_surplus(probabilities, delta, mu)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPMarket-compute_hhi"></a>}}
\if{latex}{\out{\hypertarget{method-BLPMarket-compute_hhi}{}}}
\subsection{Method \code{compute_hhi()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPMarket$compute_hhi(shares = NULL, firm_ids = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPMarket-compute_markups"></a>}}
\if{latex}{\out{\hypertarget{method-BLPMarket-compute_markups}{}}}
\subsection{Method \code{compute_markups()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPMarket$compute_markups(prices, costs)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPMarket-compute_profits"></a>}}
\if{latex}{\out{\hypertarget{method-BLPMarket-compute_profits}{}}}
\subsection{Method \code{compute_profits()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPMarket$compute_profits(prices, shares, costs)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPMarket-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BLPMarket-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPMarket$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
