% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{BLPParameters}
\alias{BLPParameters}
\title{BLP Parameter Manager}
\description{
Manages BLP model parameter compression, expansion, bounds, and labels.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BLPParameters-new}{\code{BLPParameters$new()}}
\item \href{#method-BLPParameters-compress}{\code{BLPParameters$compress()}}
\item \href{#method-BLPParameters-expand}{\code{BLPParameters$expand()}}
\item \href{#method-BLPParameters-get_bounds}{\code{BLPParameters$get_bounds()}}
\item \href{#method-BLPParameters-get_labels}{\code{BLPParameters$get_labels()}}
\item \href{#method-BLPParameters-n_free}{\code{BLPParameters$n_free()}}
\item \href{#method-BLPParameters-print}{\code{BLPParameters$print()}}
\item \href{#method-BLPParameters-clone}{\code{BLPParameters$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPParameters-new"></a>}}
\if{latex}{\out{\hypertarget{method-BLPParameters-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPParameters$new(
  sigma = NULL,
  pi = NULL,
  rho = NULL,
  beta = NULL,
  gamma = NULL,
  sigma_bounds = NULL,
  pi_bounds = NULL,
  rho_bounds = NULL,
  beta_bounds = NULL,
  gamma_bounds = NULL,
  rc_types = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPParameters-compress"></a>}}
\if{latex}{\out{\hypertarget{method-BLPParameters-compress}{}}}
\subsection{Method \code{compress()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPParameters$compress(
  sigma = NULL,
  pi = NULL,
  rho = NULL,
  beta = NULL,
  gamma = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPParameters-expand"></a>}}
\if{latex}{\out{\hypertarget{method-BLPParameters-expand}{}}}
\subsection{Method \code{expand()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPParameters$expand(theta)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPParameters-get_bounds"></a>}}
\if{latex}{\out{\hypertarget{method-BLPParameters-get_bounds}{}}}
\subsection{Method \code{get_bounds()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPParameters$get_bounds()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPParameters-get_labels"></a>}}
\if{latex}{\out{\hypertarget{method-BLPParameters-get_labels}{}}}
\subsection{Method \code{get_labels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPParameters$get_labels()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPParameters-n_free"></a>}}
\if{latex}{\out{\hypertarget{method-BLPParameters-n_free}{}}}
\subsection{Method \code{n_free()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPParameters$n_free()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPParameters-print"></a>}}
\if{latex}{\out{\hypertarget{method-BLPParameters-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPParameters$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPParameters-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BLPParameters-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPParameters$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
