% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{BLPSimulationResults}
\alias{BLPSimulationResults}
\title{BLP Simulation Results}
\description{
Results from a BLP simulation equilibrium computation.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{simulation}}{The originating BLPSimulation}

\item{\code{product_data}}{Updated product data with equilibrium prices/shares}

\item{\code{delta}}{Mean utilities at equilibrium}

\item{\code{costs}}{Marginal costs}

\item{\code{prices}}{Equilibrium prices}

\item{\code{shares}}{Equilibrium shares}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BLPSimulationResults-new}{\code{BLPSimulationResults$new()}}
\item \href{#method-BLPSimulationResults-to_problem}{\code{BLPSimulationResults$to_problem()}}
\item \href{#method-BLPSimulationResults-print}{\code{BLPSimulationResults$print()}}
\item \href{#method-BLPSimulationResults-clone}{\code{BLPSimulationResults$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPSimulationResults-new"></a>}}
\if{latex}{\out{\hypertarget{method-BLPSimulationResults-new}{}}}
\subsection{Method \code{new()}}{
Create simulation results
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPSimulationResults$new(
  simulation,
  product_data,
  delta,
  costs,
  prices,
  shares
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPSimulationResults-to_problem"></a>}}
\if{latex}{\out{\hypertarget{method-BLPSimulationResults-to_problem}{}}}
\subsection{Method \code{to_problem()}}{
Convert to a BLPProblem for estimation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPSimulationResults$to_problem(
  product_formulations = NULL,
  add_instruments = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{product_formulations}}{Optional override formulations}

\item{\code{add_instruments}}{Whether to auto-add BLP instruments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A BLPProblem object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPSimulationResults-print"></a>}}
\if{latex}{\out{\hypertarget{method-BLPSimulationResults-print}{}}}
\subsection{Method \code{print()}}{
Print simulation results
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPSimulationResults$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BLPSimulationResults-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BLPSimulationResults-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BLPSimulationResults$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
