% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construction.R
\name{build_blp_instruments}
\alias{build_blp_instruments}
\title{Build BLP Instruments (Sums of Characteristics)}
\usage{
build_blp_instruments(X, market_ids, firm_ids)
}
\arguments{
\item{X}{Matrix of exogenous product characteristics (N x K)}

\item{market_ids}{Market identifiers (length N)}

\item{firm_ids}{Firm identifiers (length N)}
}
\value{
Matrix of instruments (N x 2K): own-firm and rival columns
}
\description{
Constructs excluded instruments using the traditional BLP approach:
sums of characteristics of other own-firm products and rival products.
}
\examples{
\dontrun{
iv <- build_blp_instruments(X_exog, product_data$market_ids, product_data$firm_ids)
}
}
