% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construction.R
\name{build_differentiation_instruments}
\alias{build_differentiation_instruments}
\title{Build Differentiation Instruments (Gandhi & Houde 2020)}
\usage{
build_differentiation_instruments(
  X,
  market_ids,
  firm_ids,
  method = "local",
  interact = NULL
)
}
\arguments{
\item{X}{Matrix of product characteristics (N x K)}

\item{market_ids}{Market identifiers (length N)}

\item{firm_ids}{Firm identifiers (length N)}

\item{method}{"local" (count nearby products) or "quadratic" (sum squared distances)}

\item{interact}{Optional matrix of interaction characteristics (N x L)}
}
\value{
Matrix of instruments
}
\description{
Constructs instruments based on product proximity in characteristic space.
}
