% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construction.R
\name{build_id_data}
\alias{build_id_data}
\title{Build Balanced ID Data}
\usage{
build_id_data(T, J, F)
}
\arguments{
\item{T}{Number of markets}

\item{J}{Number of products per market}

\item{F}{Number of firms}
}
\value{
Data frame with market_ids and firm_ids
}
\description{
Creates a balanced panel of market and firm IDs for simulation.
}
