% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_ownership_matrix}
\alias{build_ownership_matrix}
\title{Build block-diagonal ownership matrix}
\usage{
build_ownership_matrix(firm_ids, market_ids)
}
\arguments{
\item{firm_ids}{Firm IDs for each product}

\item{market_ids}{Market IDs for each product}
}
\value{
Block-diagonal ownership matrix (N x N)
}
\description{
Build block-diagonal ownership matrix
}
\keyword{internal}
