% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_algebra.R
\name{commutation_matrix}
\alias{commutation_matrix}
\title{Commutation matrix}
\usage{
commutation_matrix(m, n = m)
}
\arguments{
\item{m}{Number of rows}

\item{n}{Number of columns (default = m)}
}
\value{
Commutation matrix (m\emph{n x m}n)
}
\description{
Returns K such that K \\%*\\% vec(A) = vec(t(A)) for m x n matrix A.
}
\keyword{internal}
