% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_finite_differences}
\alias{compute_finite_differences}
\title{Compute central finite differences}
\usage{
compute_finite_differences(f, x, epsilon = NULL)
}
\arguments{
\item{f}{Function taking numeric vector}

\item{x}{Point at which to evaluate}

\item{epsilon}{Step size}
}
\value{
Numeric vector of partial derivatives
}
\description{
Compute central finite differences
}
\keyword{internal}
