% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_statistics.R
\name{compute_gmm_parameter_covariances}
\alias{compute_gmm_parameter_covariances}
\title{Compute GMM Parameter Covariances}
\usage{
compute_gmm_parameter_covariances(W, S, G, se_type = "robust")
}
\arguments{
\item{W}{Weighting matrix}

\item{S}{Moment covariance matrix}

\item{G}{Moment Jacobian (d_moments / d_theta)}

\item{se_type}{"robust", "clustered", or "unadjusted"}
}
\value{
Parameter covariance matrix V
}
\description{
Compute GMM Parameter Covariances
}
\keyword{internal}
