% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_statistics.R
\name{compute_gmm_sensitivity}
\alias{compute_gmm_sensitivity}
\title{Compute Parameter Sensitivity (Andrews, Gentzkow, Shapiro 2017)}
\usage{
compute_gmm_sensitivity(W, G)
}
\arguments{
\item{W}{Weighting matrix}

\item{G}{Moment Jacobian}
}
\value{
Sensitivity matrix Lambda
}
\description{
Compute Parameter Sensitivity (Andrews, Gentzkow, Shapiro 2017)
}
\keyword{internal}
