% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_statistics.R
\name{compute_gmm_weights}
\alias{compute_gmm_weights}
\title{Compute GMM Weighting Matrix}
\usage{
compute_gmm_weights(S)
}
\arguments{
\item{S}{Moment covariance matrix}
}
\value{
Weighting matrix W = inverse of S, symmetrized
}
\description{
Compute GMM Weighting Matrix
}
\keyword{internal}
