% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_algebra.R
\name{detect_collinearity}
\alias{detect_collinearity}
\title{Detect collinearity via QR decomposition}
\usage{
detect_collinearity(x)
}
\arguments{
\item{x}{Matrix}
}
\value{
List with \code{detected} (logical) and \code{indices} (which columns to drop)
}
\description{
Detect collinearity via QR decomposition
}
\keyword{internal}
