% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_algebra.R
\name{detect_singularity}
\alias{detect_singularity}
\title{Detect near-singularity}
\usage{
detect_singularity(x)
}
\arguments{
\item{x}{Square matrix}
}
\value{
Logical: TRUE if matrix appears singular
}
\description{
Detect near-singularity
}
\keyword{internal}
