% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_algebra.R
\name{duplication_matrix}
\alias{duplication_matrix}
\title{Duplication matrix}
\usage{
duplication_matrix(n)
}
\arguments{
\item{n}{Matrix dimension}
}
\value{
Duplication matrix (n^2 x n*(n+1)/2)
}
\description{
Returns D_n such that D_n \%*\% vech(A) = vec(A) for symmetric n x n matrix A.
}
\keyword{internal}
