% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_algebra.R
\name{elimination_matrix}
\alias{elimination_matrix}
\title{Elimination matrix}
\usage{
elimination_matrix(n)
}
\arguments{
\item{n}{Matrix dimension}
}
\value{
Elimination matrix (n*(n+1)/2 x n^2)
}
\description{
Returns L_n such that L_n \%*\% vec(A) = vech(A) for symmetric n x n matrix A.
}
\keyword{internal}
