% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{find_extdata_}
\alias{find_extdata_}
\title{Find and load extdata file (works in both installed and development mode)}
\usage{
find_extdata_(filename)
}
\arguments{
\item{filename}{CSV file name}
}
\value{
Data frame
}
\description{
Find and load extdata file (works in both installed and development mode)
}
\keyword{internal}
