% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_number}
\alias{format_number}
\title{Format number for display}
\usage{
format_number(x, digits = NULL)
}
\arguments{
\item{x}{Numeric}

\item{digits}{Number of significant digits}
}
\value{
Character string
}
\description{
Format number for display
}
\keyword{internal}
