% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_seconds}
\alias{format_seconds}
\title{Format seconds into human-readable string}
\usage{
format_seconds(seconds)
}
\arguments{
\item{seconds}{Numeric}
}
\value{
Character string
}
\description{
Format seconds into human-readable string
}
\keyword{internal}
