% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{gauss_hermite}
\alias{gauss_hermite}
\title{Gauss-Hermite quadrature points and weights}
\usage{
gauss_hermite(n)
}
\arguments{
\item{n}{Number of points}
}
\value{
List with nodes and weights for integrating f(x) * exp(-x^2)
}
\description{
Gauss-Hermite quadrature points and weights
}
\keyword{internal}
