% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{groups_expand}
\alias{groups_expand}
\title{Expand group-level values to observation level}
\usage{
groups_expand(x_group, ids)
}
\arguments{
\item{x_group}{Named vector of group-level values}

\item{ids}{Group identifiers for each observation}
}
\value{
Vector expanded to observation level
}
\description{
Expand group-level values to observation level
}
\keyword{internal}
