% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{groups_mean}
\alias{groups_mean}
\title{Mean within groups}
\usage{
groups_mean(x, ids)
}
\arguments{
\item{x}{Numeric vector or matrix}

\item{ids}{Group identifiers}
}
\value{
Named vector or matrix of group means
}
\description{
Mean within groups
}
\keyword{internal}
