% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{groups_sum}
\alias{groups_sum}
\title{Sum within groups}
\usage{
groups_sum(x, ids)
}
\arguments{
\item{x}{Numeric vector or matrix}

\item{ids}{Factor or character vector of group identifiers}
}
\value{
Named vector or matrix of group sums
}
\description{
Sum within groups
}
\keyword{internal}
