% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_statistics.R
\name{iv_estimate}
\alias{iv_estimate}
\title{IV/GMM Estimation of Linear Parameters}
\usage{
iv_estimate(X, Z, W, y, jacobian = NULL)
}
\arguments{
\item{X}{Regressor matrix (N x K)}

\item{Z}{Instrument matrix (N x M)}

\item{W}{Weighting matrix (M x M)}

\item{y}{Dependent variable (N x 1)}

\item{jacobian}{Optional Jacobian of y w.r.t. nonlinear parameters (N x P)}
}
\value{
List with parameters, residuals, covariances, and residual_jacobian
}
\description{
Concentrates out linear parameters via instrumental variables.
}
\keyword{internal}
