% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_blp_agents}
\alias{load_blp_agents}
\title{Load BLP (1995) automobile agent data}
\usage{
load_blp_agents()
}
\value{
Data frame with automobile agent data
}
\description{
Load BLP (1995) automobile agent data
}
