% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_blp_products}
\alias{load_blp_products}
\title{Load BLP (1995) automobile product data}
\usage{
load_blp_products()
}
\value{
Data frame with automobile product data
}
\description{
Load BLP (1995) automobile product data
}
