% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_mixtape_demographics}
\alias{load_mixtape_demographics}
\title{Load Mixtape Sessions cereal demographic data}
\usage{
load_mixtape_demographics()
}
\value{
Data frame with Mixtape cereal demographic data
}
\description{
Load Mixtape Sessions cereal demographic data
}
