% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_mixtape_products}
\alias{load_mixtape_products}
\title{Load Mixtape Sessions cereal product data}
\usage{
load_mixtape_products()
}
\value{
Data frame with Mixtape cereal product data
}
\description{
Load Mixtape Sessions cereal product data
}
