% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_nevo_agents}
\alias{load_nevo_agents}
\title{Load Nevo (2000) cereal agent data}
\usage{
load_nevo_agents()
}
\value{
Data frame with cereal agent data
}
\description{
Load Nevo (2000) cereal agent data
}
