% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_nevo_products}
\alias{load_nevo_products}
\title{Load Nevo (2000) cereal product data}
\usage{
load_nevo_products()
}
\value{
Data frame with cereal product data
}
\description{
Load Nevo (2000) cereal product data
}
