% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_safe}
\alias{log_safe}
\title{Safe logarithm with underflow protection}
\usage{
log_safe(x, min_val = .Machine$double.xmin)
}
\arguments{
\item{x}{Numeric vector}

\item{min_val}{Minimum value to clamp to}
}
\value{
log(pmax(x, min_val))
}
\description{
Safe logarithm with underflow protection
}
\keyword{internal}
