% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_algebra.R
\name{mat_vec}
\alias{mat_vec}
\title{Vectorize a matrix (column-major)}
\usage{
mat_vec(x)
}
\arguments{
\item{x}{Matrix}
}
\value{
Numeric vector
}
\description{
Vectorize a matrix (column-major)
}
\keyword{internal}
