% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_algebra.R
\name{mat_vech}
\alias{mat_vech}
\title{Half-vectorize a symmetric matrix (lower triangle, column-major)}
\usage{
mat_vech(x)
}
\arguments{
\item{x}{Symmetric matrix}
}
\value{
Numeric vector
}
\description{
Half-vectorize a symmetric matrix (lower triangle, column-major)
}
\keyword{internal}
