% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro.R
\name{micro_dataset}
\alias{micro_dataset}
\title{Create a Micro Dataset}
\usage{
micro_dataset(name, observations, compute_weights, market_ids = NULL)
}
\arguments{
\item{name}{Dataset name}

\item{observations}{Number of observations}

\item{compute_weights}{Function(market_id, products, agents) -> I x J weights}

\item{market_ids}{Optional market subset}
}
\value{
A MicroDataset object
}
\description{
Create a Micro Dataset
}
