% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro.R
\name{micro_moment}
\alias{micro_moment}
\title{Create a Micro Moment}
\usage{
micro_moment(name, value, parts, compute_value = NULL, compute_gradient = NULL)
}
\arguments{
\item{name}{Moment name}

\item{value}{Observed target value}

\item{parts}{Single MicroPart or list of MicroPart objects}

\item{compute_value}{Function of part values -> scalar}

\item{compute_gradient}{Function of part values -> gradient}
}
\value{
A MicroMoment object
}
\description{
Create a Micro Moment
}
