% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro.R
\name{micro_part}
\alias{micro_part}
\title{Create a Micro Part}
\usage{
micro_part(name, dataset, compute_values)
}
\arguments{
\item{name}{Part name}

\item{dataset}{MicroDataset object}

\item{compute_values}{Function(market_id, products, agents) -> I x J values}
}
\value{
A MicroPart object
}
\description{
Create a Micro Part
}
