% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro.R
\name{MicroDataset}
\alias{MicroDataset}
\title{Micro Dataset}
\description{
Defines a micro-level dataset for micro moments.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Dataset name}

\item{\code{observations}}{Number of observations in micro data}

\item{\code{compute_weights}}{Function(market_id, products, agents) -> I x J weights}

\item{\code{market_ids}}{Optional subset of markets}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MicroDataset-new}{\code{MicroDataset$new()}}
\item \href{#method-MicroDataset-print}{\code{MicroDataset$print()}}
\item \href{#method-MicroDataset-clone}{\code{MicroDataset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MicroDataset-new"></a>}}
\if{latex}{\out{\hypertarget{method-MicroDataset-new}{}}}
\subsection{Method \code{new()}}{
Create micro dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MicroDataset$new(name, observations, compute_weights, market_ids = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Dataset name}

\item{\code{observations}}{Number of micro observations}

\item{\code{compute_weights}}{Function returning agent x product weight matrix}

\item{\code{market_ids}}{Optional market subset}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MicroDataset-print"></a>}}
\if{latex}{\out{\hypertarget{method-MicroDataset-print}{}}}
\subsection{Method \code{print()}}{
Print the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MicroDataset$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MicroDataset-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MicroDataset-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MicroDataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
