% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro.R
\name{MicroMoment}
\alias{MicroMoment}
\title{Micro Moment}
\description{
Matches a micro data target to a model prediction.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Moment name}

\item{\code{value}}{Observed (target) value from micro data}

\item{\code{parts}}{List of MicroPart objects}

\item{\code{compute_value_fn}}{Function of part values -> scalar}

\item{\code{compute_gradient_fn}}{Function of part values -> gradient vector}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MicroMoment-new}{\code{MicroMoment$new()}}
\item \href{#method-MicroMoment-compute_simulated_value}{\code{MicroMoment$compute_simulated_value()}}
\item \href{#method-MicroMoment-print}{\code{MicroMoment$print()}}
\item \href{#method-MicroMoment-clone}{\code{MicroMoment$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MicroMoment-new"></a>}}
\if{latex}{\out{\hypertarget{method-MicroMoment-new}{}}}
\subsection{Method \code{new()}}{
Create micro moment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MicroMoment$new(
  name,
  value,
  parts,
  compute_value = NULL,
  compute_gradient = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Moment name}

\item{\code{value}}{Observed scalar target}

\item{\code{parts}}{Single MicroPart or list of MicroPart objects}

\item{\code{compute_value}}{Function of part values -> scalar (default: identity)}

\item{\code{compute_gradient}}{Function of part values -> gradient (default: 1)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MicroMoment-compute_simulated_value"></a>}}
\if{latex}{\out{\hypertarget{method-MicroMoment-compute_simulated_value}{}}}
\subsection{Method \code{compute_simulated_value()}}{
Compute simulated moment value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MicroMoment$compute_simulated_value(
  economy,
  delta,
  sigma = NULL,
  pi = NULL,
  rho = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{economy}}{BLPEconomy object}

\item{\code{delta}}{Mean utilities}

\item{\code{sigma}}{Sigma matrix}

\item{\code{pi}}{Pi matrix}

\item{\code{rho}}{Rho vector}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Scalar simulated value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MicroMoment-print"></a>}}
\if{latex}{\out{\hypertarget{method-MicroMoment-print}{}}}
\subsection{Method \code{print()}}{
Print the moment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MicroMoment$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MicroMoment-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MicroMoment-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MicroMoment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
