% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro.R
\name{MicroPart}
\alias{MicroPart}
\title{Micro Part}
\description{
A single component of a micro moment calculation.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Part name}

\item{\code{dataset}}{MicroDataset object}

\item{\code{compute_values}}{Function(market_id, products, agents) -> I x J values}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MicroPart-new}{\code{MicroPart$new()}}
\item \href{#method-MicroPart-compute}{\code{MicroPart$compute()}}
\item \href{#method-MicroPart-print}{\code{MicroPart$print()}}
\item \href{#method-MicroPart-clone}{\code{MicroPart$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MicroPart-new"></a>}}
\if{latex}{\out{\hypertarget{method-MicroPart-new}{}}}
\subsection{Method \code{new()}}{
Create a micro part
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MicroPart$new(name, dataset, compute_values)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Part name}

\item{\code{dataset}}{MicroDataset object}

\item{\code{compute_values}}{Function returning agent x product value matrix}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MicroPart-compute"></a>}}
\if{latex}{\out{\hypertarget{method-MicroPart-compute}{}}}
\subsection{Method \code{compute()}}{
Compute aggregated part value across markets
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MicroPart$compute(economy, delta, sigma = NULL, pi = NULL, rho = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{economy}}{BLPEconomy object}

\item{\code{delta}}{Mean utilities}

\item{\code{sigma}}{Sigma matrix}

\item{\code{pi}}{Pi matrix}

\item{\code{rho}}{Rho vector}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Scalar aggregated value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MicroPart-print"></a>}}
\if{latex}{\out{\hypertarget{method-MicroPart-print}{}}}
\subsection{Method \code{print()}}{
Print the part
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MicroPart$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MicroPart-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MicroPart-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MicroPart$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
