% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mixtape_demographics}
\alias{mixtape_demographics}
\title{Mixtape Sessions Cereal Demographic Data}
\format{
A data frame with 1880 rows and the following columns:
\describe{
\item{market}{Market identifier (must match mixtape_products)}
\item{quarterly_income}{Quarterly income in dollars}
}
}
\source{
\url{https://github.com/Mixtape-Sessions/Demand-Estimation}
}
\usage{
mixtape_demographics
}
\description{
Demographic data for the Mixtape Sessions demand estimation exercises.
Contains 20 individuals per market drawn from the Current Population Survey.
}
\references{
Conlon, C. & Gortmaker, J. (2020). Best Practices for Differentiated
Products Demand Estimation with pyblp. \emph{RAND Journal of Economics},
51(4), 1108-1161.
}
\keyword{datasets}
