% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{rblp_options}
\alias{rblp_options}
\title{Get or Set rblp Package Options}
\usage{
rblp_options(...)
}
\arguments{
\item{...}{Named arguments to set options. If empty, returns all current options.}
}
\value{
Invisibly returns the previous values of changed options, or all options if none changed.
}
\description{
Get or Set rblp Package Options
}
\examples{
rblp_options(verbose = FALSE)
rblp_options()
}
