% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_exp}
\alias{safe_exp}
\title{Numerically stable exponentiation}
\usage{
safe_exp(x)
}
\arguments{
\item{x}{Numeric vector or matrix}
}
\value{
exp(x) computed with overflow protection
}
\description{
Numerically stable exponentiation
}
\keyword{internal}
