% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_algebra.R
\name{vech_to_full}
\alias{vech_to_full}
\title{Reconstruct full symmetric matrix from vech}
\usage{
vech_to_full(v, n)
}
\arguments{
\item{v}{Numeric vector (half-vectorization)}

\item{n}{Matrix dimension}
}
\value{
Symmetric matrix
}
\description{
Reconstruct full symmetric matrix from vech
}
\keyword{internal}
