% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_algebra.R
\name{vech_to_lower}
\alias{vech_to_lower}
\title{Reconstruct lower triangular matrix from vech}
\usage{
vech_to_lower(v, n)
}
\arguments{
\item{v}{Numeric vector (half-vectorization)}

\item{n}{Matrix dimension}
}
\value{
Lower triangular matrix
}
\description{
Reconstruct lower triangular matrix from vech
}
\keyword{internal}
